STD-8 STD-14
the one change should fix both.  I will need to push the changes first
and run a test to see if it worked.
Added code to the router.jspf file to make sure that a user is logged in
prior to entering the bundle.
One thing needs to be added.  Managing the redirect after a user has
attempted to log in is required