Fix mobile dorpdown view
ADMINlte bundle css defaults dropdown a tags to white.  Overroad that in
the custom css file at the 768 brake point with .kapps-menu > ul > li >
a for the kapps drop down and wrapping the i tab and span tag with a h3
in the alerts dropdown.  added a div to the broadcast alerts because the
fa-clock and time created at were overlapping the broadcast title.
remove kapp attributes and replace them with space attributes